Add. Incorporated pr review comments
