+++ .gitignore
@@ -55,7 +55,6 @@ typings/
 .next
 
 dist/
-index.html
 
 # Intellij integration
 *.iml
