+++ src/__mocks__/conceptResponse.mock.ts
@@ -1,0 +1,15 @@
+export const mockConceptResponse = {
+  results: [
+    {
+      uuid: '81d6e852-3f10-11e4-adec-0800271c1b75',
+      display: 'Consultation Note',
+      links: [
+        {
+          rel: 'self',
+          uri: 'http://localhost/openmrs/ws/rest/v1/concept/81d6e852-3f10-11e4-adec-0800271c1b75',
+          resourceAlias: 'concept',
+        },
+      ],
+    },
+  ],
+}
